+++ appserver/ejb/ejb-container/src/main/resources/org/glassfish/ejb/deployment/annotation/handlers/LocalStrings.properties
@@ -49,3 +49,5 @@ enterprise.deployment.annotation.handlers.notcompsuperclass=The annotation symbo
 enterprise.deployment.annotation.handlers.notmatchcreate=Unable to find matching Home create method for Init method {0} on bean {1}.
 enterprise.deployment.annotation.handlers.emptyEJBs=No @EJB elements in @EJBs: {0}, on {1}
 enterprise.deployment.annotation.handlers.invalidbusinessinterface=The interface {0} cannot be both a local and a remote business interface
+enterprise.deployment.annotation.handlers.sfsblifecycletxnattrtypewarn=Stateful session bean {0} lifecycle callback method {1} has transaction attribute {2} with container-managed transaction demarcation.  The transaction attribute should be either REQUIRES_NEW or NOT_SUPPORTED
+
