+++ appserver/tests/appserv-tests/devtests/ejb/ejb32/sfsb/lifecycle_cb_txattr/build.properties
@@ -1,8 +1,7 @@
-<property name="module" value="ejb-ejb32-sfsb-basic"/>
+<property name="module" value="ejb-ejb32-sfsb-lifecycle_cb_txattr"/>
 <property name="appname" value="${module}"/> 
 <property name="jndiroot" value="${appname}-ejb"/>    
 <property name="assemble" value="${build.classes.dir}/archive"/>
 <property name="se.client" value="com.acme.Client"/>
-<property name="persistence.xml" value="descriptor/persistence.xml"/>
